Handle feedback
- Split filtering domains into seperate class
